requirements.txt
langchain
langchain-google-genai
fastapi
uvicorn
pydantic
python-dotenv
